OKTA-339612 Blazor Samples > Review Haisheng suggestions and implement changes if necessary.
