+++ web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/App.razor
@@ -5,12 +5,11 @@
                 <NotAuthorized>
                     @if (!context.User.Identity.IsAuthenticated)
                     {
-                        <RedirectToLogin />
+                        <RedirectToLogin /> 
                     }
                     else
                     {
-                        <p>You are not authorized to access this resource.</p>
-                    }
+    <p>You are not authorized to access this resource.</p>}
                 </NotAuthorized>
             </AuthorizeRouteView>
         </Found>
