--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Client/Program.cs
@@ -1,40 +1,0 @@
-using System;
-using System.Net.Http;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using System.Text;
-using Microsoft.AspNetCore.Components.WebAssembly.Authentication;
-using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-
-namespace okta_blazor_web_assembly_example.Client
-{
-    public class Program
-    {
-        public static async Task Main(string[] args)
-        {
-            var builder = WebAssemblyHostBuilder.CreateDefault(args);
-            builder.RootComponents.Add<App>("app");
-
-            builder.Services.AddHttpClient("okta_blazor_web_assembly_example.ServerAPI", client => client.BaseAddress = new Uri(builder.HostEnvironment.BaseAddress))
-                .AddHttpMessageHandler<BaseAddressAuthorizationMessageHandler>();
-
-            // Supply HttpClient instances that include access tokens when making requests to the server project
-            builder.Services.AddScoped(sp => sp.GetRequiredService<IHttpClientFactory>().CreateClient("okta_blazor_web_assembly_example.ServerAPI"));
-
-            builder.Services.AddOidcAuthentication(options =>
-            {
-                // Replace the Okta placeholders with your Okta values in the appsettings.json file.
-                options.ProviderOptions.Authority = builder.Configuration.GetValue<string>("Okta:Authority");
-                options.ProviderOptions.ClientId = builder.Configuration.GetValue<string>("Okta:ClientId");;
-                options.ProviderOptions.ResponseType = "code";
-            });
-
-            builder.Services.AddApiAuthorization();
-
-            await builder.Build().RunAsync();
-        }
-    }
-}
