+++ web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Pages/Counter.razor
