--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Server/Data/ApplicationDbContext.cs
@@ -1,21 +1,0 @@
-﻿using okta_blazor_web_assembly_example.Server.Models;
-using IdentityServer4.EntityFramework.Options;
-using Microsoft.AspNetCore.ApiAuthorization.IdentityServer;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Options;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace okta_blazor_web_assembly_example.Server.Data
-{
-    public class ApplicationDbContext : ApiAuthorizationDbContext<ApplicationUser>
-    {
-        public ApplicationDbContext(
-            DbContextOptions options,
-            IOptions<OperationalStoreOptions> operationalStoreOptions) : base(options, operationalStoreOptions)
-        {
-        }
-    }
-}
