--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Server/Models/ApplicationUser.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace okta_blazor_web_assembly_example.Server.Models
-{
-    public class ApplicationUser : IdentityUser
-    {
-    }
-}
