--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Server/Pages/Shared/_Layout.cshtml
@@ -1,20 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-    <title>@ViewBag.Title</title>
-    <link href="~/css/bootstrap/bootstrap.min.css" rel="stylesheet" />
-    <link href="~/css/app.css" rel="stylesheet" />
-</head>
-
-<body>
-    <div class="main">
-        <div class="content px-4">
-            @RenderBody()
-        </div>
-    </div>
-</body>
-
-</html>
