--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Server/Properties/serviceDependencies.json
@@ -1,8 +1,0 @@
-{
-  "dependencies": {
-    "mssql1": {
-      "type": "mssql",
-      "connectionId": "DefaultConnection"
-    }
-  }
-}
