--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Server/Properties/serviceDependencies.local.json
@@ -1,8 +1,0 @@
-{
-  "dependencies": {
-    "mssql1": {
-      "type": "mssql.local",
-      "connectionId": "DefaultConnection"
-    }
-  }
-}
