--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Server/appsettings.Development.json
@@ -1,14 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "IdentityServer": {
-    "Key": {
-      "Type": "Development"
-    }
-  }
-}
