--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Server/appsettings.json
@@ -1,16 +1,0 @@
-{
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=aspnet-okta_blazor_web_assembly_example.Server-5B530D3E-6624-4832-BC0E-A0012EC3E8ED;Trusted_Connection=True;MultipleActiveResultSets=true"
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "Okta": {
-    "OktaDomain": "https://{yourOktaDomain}"
-  },
-  "AllowedHosts": "*"
-}
