+++ web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/ServerMessage.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace okta_blazor_web_assembly_example
+{
+    public class ServerMessage
+    {
+        public DateTime Date { get; set; }
+
+        public string Text { get; set; }
+    }
+
+    public class ServerMessages
+    {
+        public IList<ServerMessage> Messages { get; set; }
+    }
+}
