--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Shared/WeatherForecast.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace okta_blazor_web_assembly_example.Shared
-{
-    public class WeatherForecast
-    {
-        public DateTime Date { get; set; }
-
-        public int TemperatureC { get; set; }
-
-        public string Summary { get; set; }
-
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-    }
-}
