--- web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/Shared/okta-blazor-web-assembly-example.Shared.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-  </PropertyGroup>
-
-</Project>
