+++ web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/_Imports.razor
@@ -6,5 +6,5 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using okta_blazor_web_assembly_example.Client
-@using okta_blazor_web_assembly_example.Client.Shared
+@using okta_blazor_web_assembly_example
+@using okta_blazor_web_assembly_example.Shared
