+++ web-assembly/hosted/okta-hosted-login/okta-blazor-web-assembly-example/wwwroot/appsettings.json
@@ -2,5 +2,9 @@
   "Okta": {
     "Authority": "https://{yourOktaDomain}/oauth2/default",
     "ClientId": "{ClientId}"
+  },
+  "ServerApi": {
+    "BaseAddress": "http://localhost:8000",
+    "MessageEndpoint": "api/Messages"
   }
 }
